+++ src/views/home/homeDashboard/index.vue
@@ -1,5 +1,19 @@
 <template>
-  <div class="text-white">首頁資訊</div>
+  <div>
+    <brandFeatures />
+    <brandAdvantages />
+    <popularCourses />
+    <highlightedInstructors />
+    <learnerReviews />
+    <certificateMap />
+  </div>
 </template>
 
-<script setup lang="ts"></script>
+<script setup lang="ts">
+import brandFeatures from './pages/brandFeatures/index.vue';
+import brandAdvantages from './pages/brandAdvantages/index.vue';
+import popularCourses from './pages/popularCourses/index.vue';
+import highlightedInstructors from './pages/highlightedInstructors/index.vue';
+import learnerReviews from './pages/learnerReviews/index.vue';
+import certificateMap from './pages/certificateMap/index.vue';
+</script>
